requirements.txt
pandas
gradio
openai
openpyxl
xlsxwriter
pydantic
simplejson
python-dotenv
requests